+++ financial-connections/src/main/java/com/stripe/android/financialconnections/domain/FetchLinkAccountSession.kt
@@ -1,0 +1,55 @@
+package com.stripe.android.financialconnections.domain
+
+import com.stripe.android.financialconnections.FinancialConnectionsSheetViewModel
+import com.stripe.android.financialconnections.model.LinkAccountSession
+import com.stripe.android.financialconnections.model.LinkedAccount
+import com.stripe.android.financialconnections.model.LinkedAccountList
+import com.stripe.android.financialconnections.model.ListLinkedAccountParams
+import com.stripe.android.financialconnections.repository.FinancialConnectionsRepository
+import javax.inject.Inject
+
+internal class FetchLinkAccountSession @Inject constructor(
+    private val financialConnectionsRepository: FinancialConnectionsRepository
+) {
+
+    /**
+     * Fetches the link account session with all linked accounts via pagination
+     *
+     * @param clientSecret the link account session client secret
+     *
+     * @return LinkAccountSession with all linked accounts
+     */
+    suspend operator fun invoke(clientSecret: String): LinkAccountSession {
+        val linkAccountSession = financialConnectionsRepository.getLinkAccountSession(clientSecret)
+        if (linkAccountSession.linkedAccounts.hasMore) {
+            val accounts = mutableListOf<LinkedAccount>()
+            accounts.addAll(linkAccountSession.linkedAccounts.linkedAccounts)
+
+            var nextLinkedAccountList = financialConnectionsRepository.getLinkedAccounts(
+                ListLinkedAccountParams(clientSecret, accounts.last().id)
+            )
+            accounts.addAll(nextLinkedAccountList.linkedAccounts)
+
+            while (nextLinkedAccountList.hasMore && accounts.size < FinancialConnectionsSheetViewModel.MAX_ACCOUNTS) {
+                nextLinkedAccountList = financialConnectionsRepository.getLinkedAccounts(
+                    ListLinkedAccountParams(clientSecret, accounts.last().id)
+                )
+                accounts.addAll(nextLinkedAccountList.linkedAccounts)
+            }
+
+            return LinkAccountSession(
+                id = linkAccountSession.id,
+                clientSecret = linkAccountSession.clientSecret,
+                linkedAccounts = LinkedAccountList(
+                    linkedAccounts = accounts,
+                    hasMore = nextLinkedAccountList.hasMore,
+                    url = nextLinkedAccountList.url,
+                    count = accounts.size,
+                    totalCount = nextLinkedAccountList.totalCount
+                ),
+                livemode = linkAccountSession.livemode
+            )
+        }
+        return linkAccountSession
+    }
+}